webserver_code/requirements.txt
flask
pillow
joblib==1.3.1
numpy==1.23.0
shapely==2.0.2
hdbscan==0.8.33
umap-learn==0.5.4
opencv-python==4.7.0.72
ultralyticsplus==0.0.14
sentence-transformers==2.2.2